pyproject.toml
[project]
name = "pharmacy-states-validation"
version = "2.0.0"
description = "Validates pharmacy states of operation using multiple AI providers (OpenAI, Google Gemini) with advanced search capabilities"
authors = [
    {name = "MEDvidi Team", email = "pavzagor@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.50.0",
    "google-genai>=0.3.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = []
include = [
    "*.py",
    "*.txt",
    "*.md",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]